[Turbopack] [HMR] Allow `Option`s to be returned by `VersionedContentMap`
Building on #68698, which causes `Error`s propagated from the `VersionedContentMap` to surface as errors instead of reloading the page, allow intentionally missing values to be returned.

There are scenarios in which this is desirable and not an error, such as when the content of entries change and the page must be reloaded.

In these cases, a `Update::Total(None)` is returned, which causes the page to be reloaded.

Test Plan: CI